deduplicate colorscheme
